requirements.txt
faster-whisper>=1.0.3
pywebview>=5.1
